Create `PushDistributorManager`
